adding file for swagger and cors

+++ src/main/java/com/scm/app/controller/DivisionController.java
@@ -40,7 +40,7 @@ import com.scm.app.service.DivisionService;
 		}
 		
 		@GetMapping("/getbyid")
-		public Division getById(@RequestParam("id") Integer id) {
+		public Division getById(@RequestParam("id") Long id) {
 			return service.getById(id);
 		}
