+++ src/main/java/com/scm/app/repo/DivisionRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.scm.app.model.Division;
 
-public interface DivisionRepo extends JpaRepository<Division,Integer> {
+public interface DivisionRepo extends JpaRepository<Division,Long> {
 
 }
